Make Change
Meets criteria for test case 3: returns unspent money to the user when
selecting an item.

Any instance of a float has been replaced with an integer, because I
remembered that floats are bad.

+++ include/CoinHopper.h
@@ -11,7 +11,7 @@ class CoinHopper
         CoinHopper();
         virtual ~CoinHopper();
 
-        float insert_coin(string coin);
+        unsigned int insert_coin(string coin);
 
     protected:
